+++ pytest_cases/__init__.py
@@ -1,8 +1,8 @@
 from pytest_cases.main import cases_data, CaseData, CaseDataGetter, unfold_expected_err, extract_cases_from_module, \
-    case_name, Given, ExpectedNormal, ExpectedError, test_target, case_tags, THIS_MODULE
+    case_name, Given, ExpectedNormal, ExpectedError, test_target, case_tags, THIS_MODULE, cases_generator
 
 __all__ = ['main',
            'cases_data', 'CaseData', 'CaseDataGetter', 'unfold_expected_err', 'extract_cases_from_module',
            'case_name', 'Given', 'ExpectedNormal', 'ExpectedError',
-           'test_target', 'case_tags', 'THIS_MODULE'
+           'test_target', 'case_tags', 'THIS_MODULE', 'cases_generator'
            ]