+++ pytest_cases/main.py
@@ -1,7 +1,8 @@
 import sys
 from abc import abstractmethod, ABC
 from inspect import getmembers
-from typing import Callable, Union, Optional, Any, Tuple, List
+from itertools import product
+from typing import Callable, Union, Optional, Any, Tuple, List, Dict
 
 # noinspection PyBroadException
 try:
@@ -34,22 +35,33 @@ class CaseDataGetter(ABC):
 
 
 # Type hint for the simple functions
-DataGeneratorFunc = Callable[[], CaseData]
+CaseFunc = Callable[[], CaseData]
+
+# Type hint for generator functions
+GeneratedCaseFunc = Callable[[Any], CaseData]
 
 
 class CaseDataFromFunction(CaseDataGetter):
     """
     A CaseDataGetter relying on a function
     """
-    def __init__(self, data_generator_func: DataGeneratorFunc):
+    def __init__(self, data_generator_func: Union[CaseFunc, GeneratedCaseFunc], case_name: str = None,
+                 function_kwargs: Dict[str, Any] = None):
         """
 
         :param data_generator_func:
         """
         self.f = data_generator_func
+        self.case_name = case_name
+        if function_kwargs is None:
+            function_kwargs = dict()
+        self.function_kwargs = function_kwargs
 
     def __str__(self):
-        return self.f.__name__
+        if self.case_name is not None:
+            return self.case_name
+        else:
+            return self.f.__name__
 
     def __repr__(self):
         return "Test Case Data generator - [" + self.f.__name__ + "] - " + str(self.f)
@@ -59,7 +71,7 @@ class CaseDataFromFunction(CaseDataGetter):
         This implementation relies on the inner function to generate the dataset
         :return:
         """
-        return self.f()
+        return self.f(**self.function_kwargs)
 
 
 CASE_PREFIX = 'case_'
@@ -121,10 +133,39 @@ def extract_cases_from_module(module, filter: Any=None) -> List[CaseDataGetter]:
             #  - with the optional filter tag
             if filter is None \
                     or hasattr(f, CASE_TAGS_FIELD) and filter in getattr(f, CASE_TAGS_FIELD):
-                cases_dct[f.__code__.co_firstlineno] = f
 
-    # convert into a tuple
-    cases = [CaseDataFromFunction(cases_dct[k]) for k in sorted(cases_dct.keys())]
+                # Handle case generators
+                gen = getattr(f, GENERATOR_FIELD, False)
+                if gen:
+                    already_used_names = []
+
+                    name_template = gen[0]
+                    kwargs = gen[1]
+                    param_ids = list(kwargs.keys())
+                    all_cases_params = list(product(*kwargs.values()))
+                    nb_cases_generated = len(all_cases_params)
+                    for gen_case_id, case_params_values in enumerate(all_cases_params):
+                        # build the dictionary of parameters for the case functions
+                        gen_case_params_dct = dict(zip(param_ids, case_params_values))
+
+                        # generate the case name by applying the name template
+                        gen_case_name = name_template.format(**gen_case_params_dct)
+                        if gen_case_name in already_used_names:
+                            raise ValueError("Generated function names for generator case function {} are not unique."
+                                             " Please use all parameter names in the string format variables"
+                                             "".format(f_name))
+                        else:
+                            already_used_names.append(gen_case_name)
+
+                        # finally save the result, with an artificial floating point line number to keep order in dict
+                        gen_line_nb = f.__code__.co_firstlineno + (gen_case_id / nb_cases_generated)
+                        cases_dct[gen_line_nb] = CaseDataFromFunction(f, gen_case_name, gen_case_params_dct)
+                else:
+                    # single case
+                    cases_dct[f.__code__.co_firstlineno] = CaseDataFromFunction(f)
+
+    # convert into a tuple, taking all cases in order of appearance in the code (sort by source code line number)
+    cases = [cases_dct[k] for k in sorted(cases_dct.keys())]
 
     return cases
 
@@ -171,6 +212,12 @@ def test_target(target: Any):
     """
     A simple decorator to declare that a case function is associated with a particular target.
 
+    >>> @test_target(int)
+    >>> def case_to_test_int():
+    >>>     ...
+
+    This is actually an alias for `@case_tags(target)`, but it is a bit more readable
+
     :param target: for example a function, a class... or a string representing a function, a class...
     :return:
     """
@@ -205,36 +252,28 @@ def unfold_expected_err(expected_e: ExpectedError) -> Tuple[Optional['Type[Excep
                      "validation callable")
 
 
-def is_expected_error_instance(expected_e: ExpectedError) -> bool:
-    """
-    Returns true if the given argument is an exception instance
-
-    :param expected_e:
-    :return:
-    """
-    return issubclass(type(expected_e), Exception)
-
-
-def assert_exception_equal(e: ExpectedError, expected_e: ExpectedError):
-    """
-    Utility method to assert that a received exception matches the expected one
-
-    :param e:
-    :param expected_e:
-    :return:
-    """
-    if type(expected_e) is type:
-        # compare types
-        assert issubclass(type(e), expected_e)
-    elif isinstance(expected_e, Exception):
-        # compare instances
-        assert e == expected_e
-    elif callable(expected_e):
-        # a callable to check if ok
-        assert expected_e(e)
-    else:
-        raise TypeError("ExpectedNormal exception should either by an exception type or an exception instance, found: "
-                        + str(expected_e))
-
-
 THIS_MODULE = object()
+
+
+GENERATOR_FIELD = '__cases_generator__'
+
+
+def cases_generator(name_template, **kwargs):
+    """
+    Decorator to declare a case function as being a cases generator.
+
+    >>> @cases_generator("test with i={i}", i=range(10))
+    >>> def case_10_times(i):
+    >>>     ''' Generates 10 cases '''
+    >>>     ins = dict(a=i, b=i+1)
+    >>>     outs = i+1, i+2
+    >>>     return ins, outs, None
+
+    :return:
+    """
+
+    def cases_generator_decorator(test_func):
+        setattr(test_func, GENERATOR_FIELD, (name_template, kwargs))
+        return test_func
+
+    return cases_generator_decorator